Rearrange for single-file of event definitiuons
